FIXED: quest sharing packets with response codes and its related update fields (note: you have to stand fairly close to the quest giver player to be able receive any quests, just like he was a regular quest giver) FIXED: some quest handling related fields
Also I have commented a few things for the next generation of developers :)

+++ src/arcemu-world/QuestMgr.h
@@ -65,7 +65,7 @@ public:
 	void BuildQuestUpdateAddItem(WorldPacket* data, uint32 itemid, uint32 count);
 	void BuildQuestUpdateComplete(WorldPacket* data, Quest* qst);
 	void BuildQuestFailed(WorldPacket* data, uint32 questid);
-	void SendPushToPartyResponse(Player *plr, Player* pTarget, uint32 response);
+	void SendPushToPartyResponse(Player *plr, Player* pTarget, uint8 response);
 
 	bool OnGameObjectActivate(Player *plr, GameObject *go);
 	void OnPlayerKill(Player* plr, Creature* victim, bool IsGroupKill);
